Fix _get_feed function.
The rounds attribute is a list of bullets, so this function should be
comparing the length of these lists instead of using a number.

This logs a ton of errors, but these errors seem to be correct based on
the debugging of this method.  It looks like clips are spawned without
bullets in them.